integrations/extensions/starter-kits/mail-sending/server/requirements.txt
flask==3.0.3
python-dotenv==1.0.0